requirements.txt
flask==1.1.2
gevent==21.1.2
O365==2.0.14
PyYAML==5.4.1
